Add files via upload
Please note that LazyClassifier is intended for quick and preliminary model evaluation, and the results should be interpreted with caution. It is important to thoroughly evaluate the performance of individual classifiers using appropriate evaluation metrics and techniques before making any final conclusions about their effectiveness for your specific use case.